requirements.txt
fastapi>=0.100.0
uvicorn
pydantic>=2.0
pydantic-settings>=1.0.0
pyjwt
requests